(김항수) 파일업로드시 개수 랜덤으로 파일이 업로드 되는것 수정
CountryService
CountryController
수정

+++ test26/src/main/java/ksmart/project/test26/country/service/CountryService.java
@@ -7,6 +7,8 @@ import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.mybatis.spring.SqlSessionTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,6 +16,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.servlet.ModelAndView;
 
 import ksmart.project.test26.CountryController;
 import ksmart.project.test26.country.service.Country;
@@ -53,15 +56,7 @@ public class CountryService {
 		return getCountry;
 	}
 
-	/*
-	 * // 국가 파일 1개 조회 public CountryFile getCountryFileOne(CountryFile countryFile)
-	 * { // 매개변수 CountryFile 값 확인 logger.
-	 * debug("getCountryFileOne(CountryFile countryFile) 메서드 countryFile is {}",
-	 * countryFile); CountryFile getCountryFile =
-	 * countryDao.selectCountryFileOne(countryFile); // 리턴값 확인 logger.
-	 * debug("getCountryFileOne(CountryFile countryFile) 메서드 getCountryFile is {}",
-	 * getCountryFile); return getCountryFile; }
-	 */
+	
 
 	public void modifyCountry(Country country) {
 		// 매개변수 country 값 확인
@@ -123,7 +118,7 @@ public class CountryService {
 	}
 
 	// 파일저장
-	public void addCountry(CountryCommand countryCommand, String path) {
+	public void addCountry(CountryCommand countryCommand, String path, MultipartFile multiPartFile) {
 		logger.debug("addCountry(CountryCommand countryCommand) 메서드 countryCommand is {}", countryCommand);
 		logger.debug("addCountry(CountryCommand countryCommand) 메서드 path is {}", path);
 		Country country = new Country();
@@ -135,53 +130,56 @@ public class CountryService {
 		int countryId = countryDao.selectLastId();
 		logger.debug("int countryId = countryDao.selectLastId(); countryId is {}", countryId);
 
-		for (MultipartFile file : countryCommand.getFile()) {
-			// 1. DB에 입력...
-			CountryFile countryFile = new CountryFile();
-			UUID uuid = UUID.randomUUID();
-			String fileName = uuid.toString(); // 중복되지않은 이름 랜덤...
-			String originalName = file.getOriginalFilename();
-			int pos = originalName.lastIndexOf(".");
-			logger.debug("addCountry(CountryCommand countryCommand, String path) 메서드 pos is {}", pos);
-			// 오리지널 이름에서 마지막 .
-			String fileExt = originalName.substring(pos + 1); // 오지리널 파일 확장자
-
-			long fileSize = file.getSize(); // 오지리널 파일 사이즈
-			logger.debug("long fileSize = file.getSize() fileSize is {}", fileSize);
-
-			countryFile.setCountryId(countryId);
-			logger.debug("countryFile.setFileName(fileName) countryId is {}", countryId);
-			countryFile.setFileName(fileName);
-			logger.debug("countryFile.setFileName(fileName) fileName is {}", fileName);
-			countryFile.setFileExt(fileExt);
-			logger.debug("countryFile.setFileExt(fileExt) fileExt is {}", fileExt);
-			countryFile.setFileSize(fileSize);
-			logger.debug("countryFile.setFileSize(fileSize) fileSize is {}", fileSize);
-
-			countryDao.insertCountryFile(countryFile);
-			// 2. 파일을 저장
-			File temp = new File(path + fileName);
-			try {
-				file.transferTo(temp);
-			} catch (IllegalStateException e) {
-				e.printStackTrace();
-				if (temp.exists()) {
-					if (temp.delete()) {
+			
+		if(!multiPartFile.isEmpty()) { // 국가 파일 저장을 할 때 아무것도 파일을 저장하지 않았는데 랜덤으로 용량0 여러가지 파일이 들어가는것 수정
+			for (MultipartFile file : countryCommand.getFile()) {
+				// 1. DB에 입력
+				CountryFile countryFile = new CountryFile();
+				UUID uuid = UUID.randomUUID();
+				String fileName = uuid.toString(); // 중복되지않은 이름 랜덤...
+				String originalName = file.getOriginalFilename();
+				int pos = originalName.lastIndexOf(".");
+				logger.debug("addCountry(CountryCommand countryCommand, String path) 메서드 pos is {}", pos);
+				// 오리지널 이름에서 마지막 .
+				String fileExt = originalName.substring(pos + 1); // 오지리널 파일 확장자
+	
+				long fileSize = file.getSize(); // 오지리널 파일 사이즈
+				logger.debug("long fileSize = file.getSize() fileSize is {}", fileSize);
+	
+				countryFile.setCountryId(countryId);
+				logger.debug("countryFile.setFileName(fileName) countryId is {}", countryId);
+				countryFile.setFileName(fileName);
+				logger.debug("countryFile.setFileName(fileName) fileName is {}", fileName);
+				countryFile.setFileExt(fileExt);
+				logger.debug("countryFile.setFileExt(fileExt) fileExt is {}", fileExt);
+				countryFile.setFileSize(fileSize);
+				logger.debug("countryFile.setFileSize(fileSize) fileSize is {}", fileSize);
+	
+				countryDao.insertCountryFile(countryFile);
+				// 2. 파일을 저장
+				File temp = new File(path + fileName);
+				try {
+					file.transferTo(temp);
+				} catch (IllegalStateException e) {
+					e.printStackTrace();
+					if (temp.exists()) {
+						if (temp.delete()) {
+							logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 성공", temp);
+						} else {
+							logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 실패", temp);
+						}
+					}
+				} catch (IOException e) {
+					e.printStackTrace();
+					if (temp.exists()) {
 						logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 성공", temp);
 					} else {
 						logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 실패", temp);
-					}
-				}
-			} catch (IOException e) {
-				e.printStackTrace();
-				if (temp.exists()) {
-					logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 성공", temp);
-				} else {
-					logger.debug("addCountry(CountryCommand countryCommand) 메서드 {} 파일 삭제 실패", temp);
 				}
 			}
 		}
 	}
+}
 
 	public CountryAndCountryFile getCountryAndCountryFile(int countryId) {
 		logger.debug("getCountryAndCountryFile(int countryId)");
@@ -189,6 +187,7 @@ public class CountryService {
 		return countryAndCountryFile;
 	}
 
+	// 국가 삭제할 때 해당 국가에 같이 업로드된 파일도 함께 삭제
 	public void deleteCountry(int countryId, String realPath) {
 		// country_file 테이블 삭제 및 파일삭제후 country 테이블 삭제
 		// 매개변수 country 값 확인
@@ -221,5 +220,35 @@ public class CountryService {
 		// 국가 삭제 메서드 호출
 		countryDao.deleteCountry(countryId);
 	}
-
+	
+	public ModelAndView countryFileDownload(HttpServletRequest request,String path, String fileName, String fileExt) {
+ 		logger.debug("countryFileDownload(String path, String fileName) 메서드 실행 paht is {}", path);
+ 		logger.debug("countryFileDownload(String path, String fileName) 메서드 실행 fileName is {}", fileName);
+ 		countryDao.countryFileDownload();
+ 		// 경로 + 다운받을 파일이름을 file에 입력
+ 		File file = new File(path+fileName);
+ 		logger.debug("countryFileDownload(String path, String fileName) 메서드 실행 file is {}", file);
+ 		// 입력한 파일을 읽을수 없다면 if블록실행
+ 		if(!file.canRead()) {
+ 			logger.debug("{} 파일을 찾지 못했습니다.",file);
+ 			return new ModelAndView("fileDownloadView", "file",file);
+ 		}
+ 		// 원본 파일명 request에 셋팅
+ 		request.setAttribute("fileName", fileName);
+ 		// 확장자명을 더해서 파일 셋팅
+ 		File reFile = new File(path+fileName+"."+fileExt);
+ 		logger.debug("countryFileDownload(String path, String fileName) 메서드 실행 reFile is {}", reFile);
+ 		try {
+ 			// 파일이 있다면 if블록실행
+ 			if(file.exists()) {
+ 				// 확장자명을 더한 파일명으로 수정
+ 				file.renameTo(reFile);
+ 				// 확장자명을 더한 파일을 request에 셋팅
+ 				request.setAttribute("reFile", reFile);
+ 			}
+ 		}catch(Exception e) {
+ 			e.printStackTrace();
+ 		}
+ 		return new ModelAndView("fileDownloadView", "file",reFile);
+ 	}
 }
